uno: split game into separate file
